requirements.txt
# Koudrs Tracking API - Requirements
# Generated from pyproject.toml for compatibility

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Data validation
pydantic>=2.5.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Anti-bot bypass (for AFKL Cargo)
scrapling[fetchers]>=0.4.0

# OCR (for carriers with image captchas)
pytesseract>=0.3.10
Pillow>=10.0.0

# Email notifications
resend>=2.0.0
python-dotenv>=1.0.0
